Extract catalog draft foundation modules
